enemy movement

+++ Assets/Scripts/Entity/Movement.cs
@@ -32,5 +32,5 @@ public abstract class Movement : MonoBehaviour
 
 	protected abstract float GetDirection();
 
-	protected abstract void Flip(int dir);
+	protected abstract void Flip(float dir);
 }
